1.修改问题:在注册页面长按删除功能,当页面ui未更新,再次进行删除时报错,已添加非空判断

+++ app/src/main/java/android_serialport_api/xingbang/db/GreenDaoMaster.java
@@ -361,12 +361,14 @@ public class GreenDaoMaster {
     /**
      * 删除某一发雷管
      */
-    public void deleteDetonator(String shell) {
+    public void deleteDetonator(String gkm) {
         DenatorBaseinfo entity = mDeantorBaseDao
                 .queryBuilder()
-                .where(DenatorBaseinfoDao.Properties.ShellBlastNo.eq(shell))
+                .where(DenatorBaseinfoDao.Properties.ShellBlastNo.eq(gkm))
                 .unique();
-        mDeantorBaseDao.delete(entity);
+        if(entity!=null){
+            mDeantorBaseDao.delete(entity);
+        }
     }
